Minor updates to event picker and swag purchases
This commit adds in the default event of cfsummit instead of having a
pop-up selection.

This commit also adds in a limit for purchasing swags. The limits are
defined in the backend.

This commit also fixes some of the scaling of images. Some were scale to
fit which stretches the images. This commit makes it to scale aspect fit

This commit also fixes booklet controller when user changes events.
Previously when changing events, booklet pages are created while the
previous pages are still behind the new one. This commit makes sure it
deletes those subviews.

Signed-off-by: AnthonyAmanse <ghieamanse@gmail.com>